+++ BinanceExchange.API/Models/Response/ExchangeInfoSymbolFilter.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using BinanceExchange.API.Converter;
+using BinanceExchange.API.Models.Response.Interfaces;
+
+namespace BinanceExchange.API.Models.Response
+{
+    [DataContract]
+    public class ExchangeInfoSymbolFilter
+    {
+        [DataMember(Order = 1)]
+        public string FilterType { get; private set; }
+    }
+}
