+++ BinanceExchange.API/Models/Response/ExchangeInfoSymbolFilterPrice.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using BinanceExchange.API.Converter;
+using BinanceExchange.API.Models.Response.Interfaces;
+
+namespace BinanceExchange.API.Models.Response
+{
+    [DataContract]
+    public class ExchangeInfoSymbolFilterPrice : ExchangeInfoSymbolFilter
+    {
+        [DataMember(Order = 1)]
+        public Decimal MinPrice { get; private set; }
+
+        [DataMember(Order = 2)]
+        public Decimal MaxPrice { get; private set; }
+
+        [DataMember(Order = 3)]
+        public Decimal TickSize { get; private set; }
+    }
+}
